여행 상세 조회 시, 여정, 댓글, 좋아요 수 조회

+++ src/main/java/kr/co/fastcampus/travel/domain/trip/service/dto/response/TripInfoDto.java
@@ -11,7 +11,7 @@ public record TripInfoDto(
     LocalDate startDate,
     LocalDate endDate,
     boolean isForeign,
-    Long likeCount
+    int likeCount
 ) {
 
     public static TripInfoDto from(Trip trip) {
